refactor(pipeline): extract feedback.go from engine.go [#286]
Move feedback extraction, summarization, and prior-cycle collection
logic into a dedicated feedback.go file to reduce engine.go size and
improve cohesion. Corresponding unit tests moved to feedback_test.go.

Extracted functions: extractFeedbackFrom, extractVerifyFeedback,
extractReviewFeedback, readSnippet, collectPriorReviewCycles,
collectPriorVerifyCycles, summarizeReviewFindings,
summarizeVerifyFailures, truncateLines.
